+++ tests/schema.py
@@ -150,6 +150,38 @@ class TestModelDDL(ModelDatabaseTestCase):
              'FOREIGN KEY ("user_id") REFERENCES "foo"."user" ("id"))'),
             ('CREATE INDEX "bar"."tweet_user_id" ON "tweet" ("user_id")')])
 
+    def test_model_sequence_schema(self):
+        class Ticket(TestModel):
+            value = IntegerField(sequence='tkt_seq')
+            class Meta:
+                database = self.database
+                schema = 'foo'
+
+        self.assertCreateTable(Ticket, [
+            ('CREATE TABLE "foo"."ticket" ("id" INTEGER NOT NULL PRIMARY KEY, '
+             '"value" INTEGER NOT NULL DEFAULT NEXTVAL(\'foo.tkt_seq\'))')])
+
+        # The column default has to name the sequence CREATE SEQUENCE makes.
+        ctx = self.database.get_sql_context()
+        seq = Ticket._schema._sequence_for_field(Ticket.value)
+        self.assertEqual(ctx.sql(seq).query()[0], '"foo"."tkt_seq"')
+
+    def test_model_sequence_qualified(self):
+        class Coupon(TestModel):
+            value = IntegerField(sequence='other.cpn_seq')
+            class Meta:
+                database = self.database
+                schema = 'foo'
+
+        # A sequence the user qualified is used as-is.
+        self.assertCreateTable(Coupon, [
+            ('CREATE TABLE "foo"."coupon" ("id" INTEGER NOT NULL PRIMARY KEY, '
+             '"value" INTEGER NOT NULL DEFAULT NEXTVAL(\'other.cpn_seq\'))')])
+
+        ctx = self.database.get_sql_context()
+        seq = Coupon._schema._sequence_for_field(Coupon.value)
+        self.assertEqual(ctx.sql(seq).query()[0], '"other"."cpn_seq"')
+
     def test_bigauto_and_fk(self):
         class CustomDB(SqliteDatabase):
             field_types = {